Samples use @SpringBootApplication
@SpringBootApplication enables component scanning (which was not enabled before). To prevent component scanning from picking up Spring Boot Admin's auto-configuration classes, the package of the samples had to be changed.
