+++ spring-boot-admin-samples/spring-boot-admin-sample-consul/docker-compose.yml
@@ -1,5 +1,5 @@
 services:
   consul:
-    image: library/consul
+    image: hashicorp/consul
     ports:
       - "8500:8500"
