+++ spring-boot-admin-samples/spring-boot-admin-sample-consul/src/main/java/de/codecentric/boot/admin/sample/SpringBootAdminConsulApplication.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package de.codecentric.boot.admin;
+package de.codecentric.boot.admin.sample;
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -33,24 +33,23 @@ import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 import de.codecentric.boot.admin.server.config.AdminServerProperties;
 import de.codecentric.boot.admin.server.config.EnableAdminServer;
 
-@Configuration(proxyBeanMethods = false)
-@EnableAutoConfiguration
+@SpringBootApplication
 @EnableDiscoveryClient
 @EnableAdminServer
-public class SpringBootAdminZookeeperApplication {
+public class SpringBootAdminConsulApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(SpringBootAdminZookeeperApplication.class, args);
+		SpringApplication.run(SpringBootAdminConsulApplication.class, args);
 	}
 
 	@Profile("insecure")
 	@Configuration(proxyBeanMethods = false)
 	public static class SecurityPermitAllConfig {
 
-		private final AdminServerProperties adminServer;
+		private final String adminContextPath;
 
-		public SecurityPermitAllConfig(AdminServerProperties adminServer) {
-			this.adminServer = adminServer;
+		public SecurityPermitAllConfig(AdminServerProperties adminServerProperties) {
+			this.adminContextPath = adminServerProperties.getContextPath();
 		}
 
 		@Bean
@@ -58,11 +57,10 @@ public class SpringBootAdminZookeeperApplication {
 			http.authorizeHttpRequests((authorizeRequests) -> authorizeRequests.anyRequest().permitAll())
 				.csrf((csrf) -> csrf.csrfTokenRepository(CookieCsrfTokenRepository.withHttpOnlyFalse())
 					.ignoringRequestMatchers(
-							new AntPathRequestMatcher(this.adminServer.path("/instances"), HttpMethod.POST.toString()),
-							new AntPathRequestMatcher(this.adminServer.path("/instances/*"),
+							new AntPathRequestMatcher(this.adminContextPath + "/instances", HttpMethod.POST.toString()),
+							new AntPathRequestMatcher(this.adminContextPath + "/instances/*",
 									HttpMethod.DELETE.toString()),
-							new AntPathRequestMatcher(this.adminServer.path("/actuator/**"))));
-
+							new AntPathRequestMatcher(this.adminContextPath + "/actuator/**")));
 			return http.build();
 		}
 
@@ -72,35 +70,35 @@ public class SpringBootAdminZookeeperApplication {
 	@Configuration(proxyBeanMethods = false)
 	public static class SecuritySecureConfig {
 
-		private final AdminServerProperties adminServer;
+		private final String adminContextPath;
 
-		public SecuritySecureConfig(AdminServerProperties adminServer) {
-			this.adminServer = adminServer;
+		public SecuritySecureConfig(AdminServerProperties adminServerProperties) {
+			this.adminContextPath = adminServerProperties.getContextPath();
 		}
 
 		@Bean
 		protected SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
 			SavedRequestAwareAuthenticationSuccessHandler successHandler = new SavedRequestAwareAuthenticationSuccessHandler();
 			successHandler.setTargetUrlParameter("redirectTo");
-			successHandler.setDefaultTargetUrl(this.adminServer.path("/"));
+			successHandler.setDefaultTargetUrl(this.adminContextPath + "/");
 
 			http.authorizeHttpRequests((authorizeRequests) -> authorizeRequests
-				.requestMatchers(new AntPathRequestMatcher(this.adminServer.path("/assets/**")))
+				.requestMatchers(new AntPathRequestMatcher(this.adminContextPath + "/assets/**"))
 				.permitAll()
-				.requestMatchers(new AntPathRequestMatcher(this.adminServer.path("/login")))
+				.requestMatchers(new AntPathRequestMatcher(this.adminContextPath + "/login"))
 				.permitAll()
 				.anyRequest()
 				.authenticated())
-				.formLogin((formLogin) -> formLogin.loginPage(this.adminServer.path("/login"))
+				.formLogin((formLogin) -> formLogin.loginPage(this.adminContextPath + "/login")
 					.successHandler(successHandler))
-				.logout((logout) -> logout.logoutUrl(this.adminServer.path("/logout")))
+				.logout((logout) -> logout.logoutUrl(this.adminContextPath + "/logout"))
 				.httpBasic(Customizer.withDefaults())
 				.csrf((csrf) -> csrf.csrfTokenRepository(CookieCsrfTokenRepository.withHttpOnlyFalse())
 					.ignoringRequestMatchers(
-							new AntPathRequestMatcher(this.adminServer.path("/instances"), HttpMethod.POST.toString()),
-							new AntPathRequestMatcher(this.adminServer.path("/instances/*"),
+							new AntPathRequestMatcher(this.adminContextPath + "/instances", HttpMethod.POST.toString()),
+							new AntPathRequestMatcher(this.adminContextPath + "/instances/*",
 									HttpMethod.DELETE.toString()),
-							new AntPathRequestMatcher(this.adminServer.path("/actuator/**"))));
+							new AntPathRequestMatcher(this.adminContextPath + "/actuator/**")));
 
 			return http.build();
 		}
