+++ spring-boot-admin-samples/spring-boot-admin-sample-zookeeper/docker-compose.yml
@@ -1,0 +1,5 @@
+services:
+  zookeeper:
+    image: zookeeper
+    ports:
+      - 2181:2181
